+++ packages/components/src/components/List/components/Items/views/ListBoxItem/view.ts
@@ -1,0 +1,10 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import type { ListBoxItem } from "./ListBoxItem";
+import type { ViewComponent } from "@/lib/viewComponentContext";
+
+declare global {
+  interface FlowViewComponents {
+    ItemsListBoxItem: ViewComponent<typeof ListBoxItem>;
+  }
+}