+++ packages/components/src/components/List/model/filter/types.ts
@@ -21,15 +21,10 @@ export interface FilterShape<T, TProp extends PropertyName<T>, TMatcherValue> {
   values?: readonly TMatcherValue[];
   name?: string;
   defaultSelected?: readonly NonNullable<TMatcherValue>[];
-  /**
-   * If provided, the filter will be initialized with these values instead of
-   * the defaultSelected and restored selections from any provided settings
-   * storage. It behaves like an initial user initiated selection.
-   */
-  initialSelected?: readonly NonNullable<TMatcherValue>[];
   onChange?: FilterUpdatedCallback;
   priority?: "primary" | "secondary";
   autosave?: boolean;
+  manualSave?: boolean;
   dateRangeOptions?: RangeCalendarProps;
 }