+++ packages/components/src/components/List/model/search/types.ts
@@ -17,4 +17,5 @@ export interface SearchShape<IgnoredT> {
   render?: SearchFieldRenderComponent;
   textFieldProps: SupportedSearchFieldProps;
   defaultValue?: string;
+  autosave?: boolean;
 }