+++ packages/components/src/components/List/stories/Default.stories.tsx
@@ -70,7 +70,6 @@ const meta: Meta<typeof List> = {
           <DomainList.List
             batchSize={5}
             aria-label="Domains"
-            onAction={(domain) => console.log(domain.hostname)}
             settingStorageKey="domains"
             defaultViewMode="tiles"
           >
@@ -87,7 +86,9 @@ const meta: Meta<typeof List> = {
               values={availableTypes}
               property="type"
               name="Type"
-              initialSelected={["Subdomain"]}
+              autosave
+              manualSave
+              defaultSelected={["Domain"]}
             />
 
             <DomainList.Search autoFocus />
@@ -132,7 +133,11 @@ const meta: Meta<typeof List> = {
               </DomainList.TableBody>
             </DomainList.Table>
 
-            <DomainList.Item showTiles textValue={(domain) => domain.hostname}>
+            <DomainList.Item
+              href={() => "#"}
+              showTiles
+              textValue={(domain) => domain.hostname}
+            >
               {(domain) => (
                 <ListItemView>
                   <Avatar color={domain.type === "Domain" ? "blue" : "teal"}>