+++ packages/components/src/components/Separator/Separator.tsx
@@ -4,13 +4,17 @@ import clsx from "clsx";
 import styles from "./Separator.module.scss";
 import * as Aria from "react-aria-components";
 
-export type SeparatorProps = Omit<Aria.SeparatorProps, "orientation">;
+export type SeparatorProps = Aria.SeparatorProps;
 
 /** @flr-generate all */
 export const Separator: FC<SeparatorProps> = (props) => {
-  const { className, ...rest } = props;
+  const { className, orientation = "horizontal", ...rest } = props;
 
-  const rootClassName = clsx(styles.separator, className);
+  const rootClassName = clsx(
+    styles.separator,
+    orientation === "vertical" && styles.vertical,
+    className,
+  );
 
   return (
     <Aria.Separator elementType="hr" className={rootClassName} {...rest} />