+++ packages/components/src/integrations/react-hook-form/components/Field/stories/CodeEditor.stories.tsx
@@ -1,0 +1,139 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { useEffect } from "react";
+import { useForm } from "react-hook-form";
+import { action } from "storybook/actions";
+import { Label } from "@/components/Label";
+import {
+  Field,
+  Form,
+  ResetButton,
+  SubmitButton,
+  typedField,
+} from "@/integrations/react-hook-form";
+import { Button } from "@/components/Button";
+import { Section } from "@/components/Section";
+import { ActionGroup } from "@/components/ActionGroup";
+import { sleep } from "@/lib/promises/sleep";
+import { FieldError } from "@/components/FieldError";
+import { CodeEditor } from "@/components/CodeEditor";
+import { validator as DotEnvValidator } from "@/components/CodeEditor/languages/dotEnv";
+
+const submitAction = action("submit");
+
+const meta: Meta<typeof Field> = {
+  title: "Integrations/React Hook Form/CodeEditor",
+  component: Field,
+  render: () => {
+    interface Values {
+      code: string;
+    }
+
+    const handleSubmit = async (values: Values) => {
+      await sleep(1500);
+      submitAction(values);
+    };
+
+    const form = useForm<Values>({
+      defaultValues: {
+        code: "foo=bar\nfoo=baz",
+      },
+    });
+
+    const Field = typedField(form);
+
+    return (
+      <Form form={form} onSubmit={handleSubmit}>
+        <Section>
+          <Field
+            name="code"
+            rules={{
+              required: true,
+              validate: (value) =>
+                DotEnvValidator(value) ? true : "invalid dot env",
+            }}
+          >
+            <CodeEditor language="dotEnv">
+              <Label>DotEnv</Label>
+            </CodeEditor>
+          </Field>
+
+          <ActionGroup>
+            <ResetButton>Reset</ResetButton>
+            <SubmitButton>Submit</SubmitButton>
+          </ActionGroup>
+        </Section>
+      </Form>
+    );
+  },
+};
+export default meta;
+
+type Story = StoryObj<typeof Field>;
+
+export const Default: Story = {};
+
+export const WithFieldError: Story = {
+  render: () => {
+    const form = useForm({
+      defaultValues: {
+        field: "",
+      },
+    });
+    useEffect(() => {
+      form.setError("field", {
+        type: "required",
+        message: "ErrorFromForm",
+      });
+    }, []);
+
+    return (
+      <Form form={form} onSubmit={async () => await sleep(2000)}>
+        <Field name="field">
+          <CodeEditor>
+            <Label>Code</Label>
+          </CodeEditor>
+        </Field>
+        <CodeEditor isInvalid>
+          <Label>Code</Label>
+          <FieldError>ErrorFromOuterFieldError!</FieldError>
+        </CodeEditor>
+      </Form>
+    );
+  },
+};
+
+export const WithFocus: Story = {
+  render: () => {
+    const form = useForm({
+      defaultValues: {
+        field: "",
+      },
+    });
+    return (
+      <Form form={form} onSubmit={async () => await sleep(2000)}>
+        <Field name={"field"}>
+          <CodeEditor>
+            <Label>Code</Label>
+          </CodeEditor>
+        </Field>
+        <div style={{ marginBottom: "2200px" }} />
+        <Button
+          onPress={() =>
+            form.setError(
+              "field",
+              { type: "required", message: "oh no" },
+              { shouldFocus: true },
+            )
+          }
+        >
+          err through form
+        </Button>
+        <Button onPress={() => form.setFocus("field")}>
+          focus through form
+        </Button>
+        <ResetButton>Reset</ResetButton>
+        <SubmitButton>Submit</SubmitButton>
+      </Form>
+    );
+  },
+};