+++ packages/remote-elements/src/auto-generated/RemoteItemsListBoxElement.ts
@@ -1,0 +1,132 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import { FlowRemoteElement } from "@/lib/FlowRemoteElement";
+import type { ListBoxProps as RemoteItemsListBoxElementProps } from "@mittwald/flow-react-components";
+export type { ListBoxProps as RemoteItemsListBoxElementProps } from "@mittwald/flow-react-components";
+
+export class RemoteItemsListBoxElement extends FlowRemoteElement<RemoteItemsListBoxElementProps> {
+  static override get remoteAttributes() {
+    return ["style"];
+  }
+
+  static override get remoteProperties() {
+    return {
+      "aria-describedby": {},
+      "aria-details": {},
+      "aria-label": {},
+      "aria-labelledby": {},
+      autoFocus: {},
+      className: {},
+      defaultSelectedKeys: {},
+      dependencies: {},
+      dir: {},
+      disabledKeys: {},
+      disallowEmptySelection: {},
+      dragAndDropHooks: {},
+      escapeKeyBehavior: {},
+      hidden: {},
+      id: {},
+      inert: {},
+      items: {},
+      lang: {},
+      layout: {},
+      orientation: {},
+      renderEmptyState: {},
+      selectedKeys: {},
+      selectionBehavior: {},
+      selectionMode: {},
+      shouldFocusOnHover: {},
+      shouldFocusWrap: {},
+      shouldSelectOnPressUp: {},
+      slot: {},
+      tileMaxWidth: {},
+      translate: {},
+    };
+  }
+
+  static override get remoteEvents() {
+    return {
+      action: {},
+      animationEnd: {},
+      animationEndCapture: {},
+      animationIteration: {},
+      animationIterationCapture: {},
+      animationStart: {},
+      animationStartCapture: {},
+      auxClick: {},
+      auxClickCapture: {},
+      blur: {},
+      click: {},
+      clickCapture: {},
+      contextMenu: {},
+      contextMenuCapture: {},
+      doubleClick: {},
+      doubleClickCapture: {},
+      focus: {},
+      focusChange: {},
+      gotPointerCapture: {},
+      gotPointerCaptureCapture: {},
+      lostPointerCapture: {},
+      lostPointerCaptureCapture: {},
+      mouseDown: {},
+      mouseDownCapture: {},
+      mouseEnter: {},
+      mouseLeave: {},
+      mouseMove: {},
+      mouseMoveCapture: {},
+      mouseOut: {},
+      mouseOutCapture: {},
+      mouseOver: {},
+      mouseOverCapture: {},
+      mouseUp: {},
+      mouseUpCapture: {},
+      pointerCancel: {},
+      pointerCancelCapture: {},
+      pointerDown: {},
+      pointerDownCapture: {},
+      pointerEnter: {},
+      pointerLeave: {},
+      pointerMove: {},
+      pointerMoveCapture: {},
+      pointerOut: {},
+      pointerOutCapture: {},
+      pointerOver: {},
+      pointerOverCapture: {},
+      pointerUp: {},
+      pointerUpCapture: {},
+      scroll: {},
+      scrollCapture: {},
+      selectionChange: {},
+      touchCancel: {},
+      touchCancelCapture: {},
+      touchEnd: {},
+      touchEndCapture: {},
+      touchMove: {},
+      touchMoveCapture: {},
+      touchStart: {},
+      touchStartCapture: {},
+      transitionCancel: {},
+      transitionCancelCapture: {},
+      transitionEnd: {},
+      transitionEndCapture: {},
+      transitionRun: {},
+      transitionRunCapture: {},
+      transitionStart: {},
+      transitionStartCapture: {},
+      wheel: {},
+      wheelCapture: {},
+    };
+  }
+
+  static override get remoteSlots() {
+    return [];
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    "flr-items-list-box": InstanceType<typeof RemoteItemsListBoxElement>;
+  }
+}
+
+customElements.define("flr-items-list-box", RemoteItemsListBoxElement);