+++ packages/remote-react-components/e2e/tests/Form.browser.test.remote.tsx
@@ -7,7 +7,6 @@ import {
   Option,
   Select,
   Button,
-  CodeBlock,
 } from "../../src/auto-generated";
 import { Form } from "../../src/components/Form";
 import { type FC, useState } from "react";
@@ -73,9 +72,9 @@ const TestForm: FC<{
           Submit
         </Button>
         {event && (
-          <CodeBlock
+          <TextField
             data-testid="form-result"
-            code={JSON.stringify(event, undefined, 2)}
+            value={btoa(JSON.stringify(event))}
           />
         )}
       </Section>