+++ packages/remote-react-components/src/auto-generated/CodeEditor.ts
@@ -1,0 +1,18 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+"use client";
+import createFlowRemoteComponent from "@/components/createFlowRemoteComponent";
+import { RemoteCodeEditorElement } from "@mittwald/flow-remote-elements";
+export { type RemoteCodeEditorElement } from "@mittwald/flow-remote-elements";
+
+export const CodeEditor = createFlowRemoteComponent(
+  "flr-code-editor",
+  "CodeEditor",
+  RemoteCodeEditorElement,
+  {
+    slotProps: {
+      wrapper: "flr-slot-root-wrapper",
+    },
+    eventProps: {},
+  },
+);